pyproject.toml
[tool.poetry]
name = "learning-langchain"
version = "0.1.0"
description = ""
authors = ["RutamBhagat <rutambhagat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.16"
python-dotenv = "^1.0.1"
openai = "^1.23.2"
langchain-openai = "^0.1.3"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
